+++ site/dat/docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.3.8 (next)
   - fixed bug when old jars were not removed during JMeter installation
+  - add `project` to BlazeMeter report config, allowing to sort tests by projects
 
 ## 0.3.7
   - fail criteria without timeframe is checked at the end of the test